+++ src/main/server/services/death.ts
@@ -1,5 +1,5 @@
 import * as alt from 'alt-server';
-import { useServices } from './index.js';
+import { useServiceRegister } from './index.js';
 
 export interface DeathService {
     /**
@@ -19,7 +19,7 @@ export interface DeathService {
 
 declare global {
     interface RebarServices {
-        deathService: Partial<DeathService>;
+        deathService: DeathService;
     }
 }
 
@@ -33,32 +33,16 @@ declare module 'alt-server' {
 export function useDeathService() {
     return {
         respawn(...args: Parameters<DeathService['respawn']>) {
-            const services = useServices().get('deathService');
-            if (services.length <= 0) {
-                return;
-            }
-
-            for (let service of services) {
-                if (typeof service.respawn !== 'function') {
-                    continue;
-                }
-
+            const service = useServiceRegister().get('deathService');
+            if (service && service.respawn) {
                 service.respawn(...args);
             }
 
             alt.emit('playerRespawn', ...args);
         },
         revive(...args: Parameters<DeathService['revive']>) {
-            const services = useServices().get('deathService');
-            if (services.length <= 0) {
-                return;
-            }
-
-            for (let service of services) {
-                if (typeof service.revive !== 'function') {
-                    continue;
-                }
-
+            const service = useServiceRegister().get('deathService');
+            if (service && service.respawn) {
                 service.revive(...args);
             }